+++ packages/gov-ui-kit/CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   Implement `ProposalDataListItemSkeleton` module component
 
+### Changed
+
+-   Remove padding from `DataListContainer`, `DataListFilterStatus`, `DataListPagination` and `DataListRoot`
+
 ## [1.0.23] - 2024-04-18
 
 ### Added